requirements.txt
Django~=5.0.6
djangorestframework~=3.15.1
djangorestframework-simplejwt~=5.3.1
django-extensions~=3.2.3
